refactor: folder page -> folder modal
