--- snapsplit/src/app/trip/[tripId]/snap/folder/[name]/page.tsx
@@ -1,5 +1,0 @@
-import SnapFolderPage from '@/features/trip/[tripId]/snap/folder/[name]/SnapFolderPage';
-
-export default function SnapFolder() {
-    return <SnapFolderPage />;
-}
