+++ snapsplit/src/features/trip/[tripId]/snap/_components/snapFolderModal/FilterBar.tsx
