+++ snapsplit/src/features/trip/[tripId]/snap/_components/snapFolderModal/SnapFolderHeader.tsx
