+++ snapsplit/src/features/trip/[tripId]/snap/_components/snapFolderModal/SnapFolderInfo.tsx
